requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
mahotas>=1.4.11
scikit-learn>=1.0
matplotlib>=3.4.3
pillow>=8.4.0
tensorflow>=2.6.0
keras>=2.6.0